requirements.txt
django==2.2
djangorestframework==3.9.2
django-rest-swagger
coreapi
django_filters
